Changes in case study structuring and modification
